requirements.txt
absl-py==2.2.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.5.3
certifi==2024.12.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.4
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
environ==1.0
executing==2.1.0
extensions==0.4
filelock==3.17.0
Flask==3.0.3
Flask-Cache==0.13.1
Flask-Markdown==0.3
Flask-MySQLdb==2.0.0
flask_weasyprint==1.1.0
flatbuffers==25.2.10
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==22.0.0
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.7.4.2
keras==3.9.2
kiwisolver==1.4.7
kombu==5.5.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mysqlclient==2.2.4
namex==0.0.9
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
oscrypto==1.3.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pdfkit==1.0.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
protobuf==5.29.4
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydicom==3.0.1
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.9.0
pyparsing==3.1.4
pypdf==5.1.0
pyphen==0.17.2
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
razorpay==1.4.2
regex==2024.11.6
replicate==1.0.7
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.15.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
zopfli==0.2.3.post1
